+++ src/main/java/io/github/astrapi69/lang/model/ClassModel.java
@@ -1,0 +1,84 @@
+/**
+ * The MIT License
+ *
+ * Copyright (C) 2015 Asterios Raptis
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+ * associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute,
+ * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
+ * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package io.github.astrapi69.lang.model;
+
+import java.util.List;
+import java.util.Map;
+
+import javax.lang.model.element.ElementKind;
+import javax.lang.model.element.Modifier;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import lombok.ToString;
+import lombok.experimental.FieldDefaults;
+
+/**
+ * The class {@link ClassModel}.
+ */
+@Getter
+@Setter
+@EqualsAndHashCode(exclude = {
+		"classAnnotations" }, onlyExplicitlyIncluded = false, callSuper = false)
+@ToString(exclude = { "classAnnotations" })
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder(toBuilder = true)
+@FieldDefaults(level = AccessLevel.PRIVATE)
+public class ClassModel
+{
+
+	/** The class annotations. */
+	List<ClassModel> classAnnotations;
+
+	/** The class name. */
+	String className;
+
+	/** The extended class name. */
+	String extendedClassName;
+
+	/** The generic types. */
+	List<String> genericTypes;
+
+	/** The imports. */
+	List<String> imports;
+
+	/** The interface implementations. */
+	List<String> interfaceImplementations;
+
+	/** The kind. */
+	@Builder.Default
+	ElementKind kind = ElementKind.CLASS;
+
+	/** The methods. */
+	Map<String, MethodModel> methods;
+
+	/** The modifiers. */
+	List<Modifier> modifiers;
+
+	/** The package name. */
+	String packageName;
+}