+++ src/test/java/io/github/astrapi69/lang/ClassExtensionsTest.java
@@ -462,13 +462,13 @@ public class ClassExtensionsTest
 
 		resourcesDirPath = "io/github/astrapi69/lang";
 		urls = ClassExtensions.getResources(resourcesDirPath);
-		assertTrue(urls.size() == 3);
+		assertTrue(urls.size() == 4);
 
 		urls = ClassExtensions.getResources(resourcesDirPath, "jar");
 		assertTrue(urls.size() == 3);
 
 		urls = ClassExtensions.getResources(resourcesDirPath, "file");
-		assertTrue(urls.size() == 0);
+		assertTrue(urls.size() == 1);
 	}
 
 	/**