+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 ## Change log
 ----------------------
 
-Version 3.8
+Version 3.9
+-------------
+
+ADDED:
+
+- new model classes for hold java class fields like method, field, class and annotation
+- new model classes for hold thread related fields
+  
+  Version 3.8
 -------------
 
 ADDED: